Fix the decorator
The point of the Decorator is to decorate one object with different
behaviours. Therefore, the example should include more than one
behaviour, so I added "Filled" to "Red". Also, I fixed the name of the
class to fully correspond to its path. I deleted the Rectangle and
Square as they weren't needed to demonstrate what the decorator does.