Improve the Observer example
Convert the Observer to a role. It shortens the code and is usually
better than abstract classes in Perl.

Make the observer links in Subject weak. Add a detach method to
subject and call it from the Observers' destructor. Don't acces
attributes as hash elements, use accessors.

Test the new behaviour.