switch to MooX::Clone
This change does not alter the behaviour of the code, but the underlying mechanics. Instead of inheriting from Clone, which is not directly related to the project and the classes being used here, a role is now dynamically applied. This means the ISA of our classes stays clean, while the interface remains the same.